refactor: a little rxjs sample cleaning
