+++ src/3-rxjs/basics/basics.module.ts
@@ -28,9 +28,6 @@ const routes: Routes = [{ path: '', pathMatch: 'full', component: BasicsComponen
     Basic06Component,
     Basic07Component,
     BasicsComponent,
-  ],
-  exports: [
-    BasicsComponent
   ]
 })
 export class BasicsModule { }
