+++ src/3-rxjs/core/index.ts
@@ -1,4 +1,5 @@
 export * from './busy.service';
+export * from './helpers';
 export * from './in-memory-data.service';
 export * from './model';
 export * from './message.service';
