+++ src/3-rxjs/play-subject/play-subject.component.ts
@@ -1,11 +1,10 @@
 // #region imports
 // tslint:disable:member-ordering
 import { Component, OnDestroy } from '@angular/core';
-import { log, logOp, loggingObserver, messageObserver } from '../helpers';
+import { logOp, messageObserver } from '../core/helpers';
 
 // Always available in the subject playground
 import { Observable, Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
 
 // Namespace to get something you need
 import * as Rxjs from 'rxjs';
